Added a few options in the the easter egg function, added the audio for wring and right answer, and move almost al the js file to public/js and updated the paths in the index.js
