+++ public/js/databaseConnection.js
